pyproject.toml
[project]
name = "autosummarize"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "flask>=3.1.1",
    "google-genai>=1.28.0",
    "gunicorn>=23.0.0",
    "moviepy>=2.2.1",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "rich>=14.1.0",
    "werkzeug>=3.1.3",
]
